Fixes up file upload to work.
This required installing undici, which is the more frequently updated version of Node's built in fetch. This is because Langflow doesn't consider a multipart/form-data request that doesn't end with a CRLF to be a valid request and the current version of undici embedded in Node doesn't add the CRLF. I've made a PR to Langflow to relax this requirement (https://github.com/langflow-ai/langflow/pull/5660), and version 7 of undici has been merged to the Node main branch but not yet released in any version, at the time of writing.