Refactor baseUrl handling in LangflowClient constructor and add tests for missing baseUrl scenarios
If baseUrl is provided as an property on the opts, but is it undefined or null, this is likely not the intention. Previously the baseUrl would get set to the DATASTAX_LANGFLOW_BASE_URL and the instantiation would fail due to a missing langflowId. Now it will throw an more useful error, unless langflowId and apiKey are present.